+++ src/commands/dataconnect-sdk-generate.ts
@@ -12,6 +12,7 @@ type GenerateOptions = Options & { watch?: boolean };
 
 export const command = new Command("dataconnect:sdk:generate")
   .description("generates typed SDKs for your Data Connect connectors")
+  .firebaseNotRequired()
   .option(
     "--watch",
     "watch for changes to your connector GQL files and regenerate your SDKs when updates occur",