+++ src/commands/dataconnect-sql-grant.ts
@@ -13,6 +13,7 @@ const allowedRoles = Object.keys(fdcSqlRoleMap);
 
 export const command = new Command("dataconnect:sql:grant [serviceId]")
   .description("Grants the SQL role <role> to the provided user or service account <email>.")
+  .firebaseNotRequired()
   .option("-R, --role <role>", "The SQL role to grant. One of: owner, writer, or reader.")
   .option(
     "-E, --email <email>",