+++ src/commands/firestore-backups-get.ts
@@ -9,6 +9,7 @@ import { PrettyPrint } from "../firestore/pretty-print";
 
 export const command = new Command("firestore:backups:get <backup>")
   .description("Get a Cloud Firestore database backup.")
+  .firebaseNotRequired()
   .before(requirePermissions, ["datastore.backups.get"])
   .before(warnEmulatorNotSupported, Emulators.FIRESTORE)
   .action(async (backupName: string, options: FirestoreOptions) => {